requirements.txt
alembic==1.7.1
asgiref==3.7.2
click==8.1.7
fastapi==0.68.1
greenlet==2.0.2
h11==0.14.0
asyncpg==0.24.0
pydantic==1.10.12
python-dotenv==1.0.0
SQLAlchemy==1.4.49
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.4
starlette==0.14.2
typing_extensions==4.7.1
uvicorn==0.15.0
